requirements.txt
Flask==1.0.2
sqlalchemy==1.2.11
flask_sqlalchemy==2.3.2
flask_bcrypt==0.7.1
flask_wtf==0.14.2
flask_login==0.4.1
pillow==5.2.0
bcrypt==3.1.4
wtforms==2.2.1
fstring==1.7.4